+++ jax_cosmo/__init__.py
@@ -5,12 +5,11 @@ from __future__ import print_function
 
 from jax_cosmo.core import *
 from jax_cosmo.parameters import *
-# import jax_cosmo.background as background
-# import jax_cosmo.nonlinear as nonlinear
-# import jax_cosmo.transfer as transfer
-# import jax_cosmo.power as power
-# import jax_cosmo.redshift as redshift
-# import jax_cosmo.bias as bias
-# import jax_cosmo.probes as probes
-# import jax_cosmo.angular_cl as cl
-# import jax_cosmo.likelihood as likelihood
+import jax_cosmo.background as background
+import jax_cosmo.transfer as transfer
+import jax_cosmo.power as power
+import jax_cosmo.redshift as redshift
+import jax_cosmo.bias as bias
+import jax_cosmo.probes as probes
+import jax_cosmo.angular_cl as cl
+import jax_cosmo.likelihood as likelihood
